+++ RELEASES/RELEASES.md
@@ -1,8 +1,8 @@
 # RELEASES
 
-## V1.0.1-beta.3 (Build 20260829001)
+## V1.0.1-beta.4 (Build 20260830001)
 
-**Release date:** 2026-08-29
+**Release date:** 2026-08-30
 
 ### Features
 
@@ -12,6 +12,11 @@
 - Improved media file importing and video stabilization to reduce excessive memory and CPU usage and prevent related app crashes.
 - Enhanced the file import demo with a before-and-after view for comparing video stabilization results.
 - Added a new API to format device media storage by deleting all media files.
+- Added support for determining whether the connected device supports the find-device feature.
+- Added support for retrieving the following device information:
+  - Image-enhancement post-processing algorithm currently in use.
+  - Recommended maximum recording durations for video and audio.
+  - Minimum battery level required for taking photos, recording video or audio, and transferring files.
 - Improved SDK stability.
 
 ## V1.0.0 (Build 20260824001)